fix(vqb): refacto on aggregates
